Add files via upload
Todo list is a basic application in which users can add items.  It’s a list of tasks you need to complete or things that you want to do.  In to-do, list tasks are organized in order of priority.   One of the most important reasons you should use a to-do list is that it will help you stay organized.